requirements.txt
Flask==1.1.2
Jinja2==2.11.2
PyMySQL==1.0.2
python-dotenv==0.15.0
cryptography